Mean reversion/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mean_reversion_bot"
version = "0.1.0"
description = "Hummingbot-compatible mean reversion trading service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pybit==5.9.0",
  "asyncpg==0.29.0",
  "redis==4.5.1",
  "structlog==23.3.0",
  "python-dotenv==1.0.1",
  "pandas==2.2.3",
  "ta==0.10.1",
  "aiohttp==3.9.3",
  "websockets==12.0",
  "numpy==1.24.2",
  "scikit-learn==1.2.2",
  "tensorflow==2.15.0",
  "keras-tuner>=1.0.3",
  "shap>=0.41.0"
]

[project.scripts]
backtester  = "scripts.backtester:main"
monte-carlo = "scripts.monte_carlo:main"
live-tuner  = "scripts.live_tuner:monitor"
